backend/requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
Pillow==10.1.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
motor==3.3.2
aiofiles==23.2.1
